+++ test/fixtures/monkey-patch-run-main.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const oldRunMain = require('module').runMain;
+
+require('module').runMain = function(...args) {
+  console.log('runMain is monkey patched!');
+  oldRunMain(...args);
+};
