+++ test/message/events_unhandled_error_nexttick.out
@@ -8,7 +8,7 @@ Error
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:*:*)
     at Module.load (internal/modules/cjs/loader.js:*:*)
     at Function.Module._load (internal/modules/cjs/loader.js:*:*)
-    at Function.Module.runMain (internal/modules/cjs/loader.js:*:*)
+    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*:*)
     at internal/main/run_main_module.js:*:*
 Emitted 'error' event at:
     at *events_unhandled_error_nexttick.js:*:*
