+++ test/message/if-error-has-good-stack.out
@@ -15,7 +15,7 @@ AssertionError [ERR_ASSERTION]: ifError got unwanted exception: test error
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:*:*)
     at Module.load (internal/modules/cjs/loader.js:*:*)
     at Function.Module._load (internal/modules/cjs/loader.js:*:*)
-    at Function.Module.runMain (internal/modules/cjs/loader.js:*:*)
+    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*:*)
     at internal/main/run_main_module.js:*:* {
   generatedMessage: false,
   code: 'ERR_ASSERTION',
@@ -28,7 +28,7 @@ AssertionError [ERR_ASSERTION]: ifError got unwanted exception: test error
       at Object.Module._extensions..js (internal/modules/cjs/loader.js:*:*)
       at Module.load (internal/modules/cjs/loader.js:*:*)
       at Function.Module._load (internal/modules/cjs/loader.js:*:*)
-      at Function.Module.runMain (internal/modules/cjs/loader.js:*:*)
+      at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*:*)
       at internal/main/run_main_module.js:*:*
   expected: null,
   operator: 'ifError'
