+++ test/message/vm_display_runtime_error.out
@@ -12,7 +12,7 @@ Error: boo!
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:*)
     at Module.load (internal/modules/cjs/loader.js:*)
     at Function.Module._load (internal/modules/cjs/loader.js:*)
-    at Function.Module.runMain (internal/modules/cjs/loader.js:*)
+    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*)
     at internal/main/run_main_module.js:*:*
 test.vm:1
 throw new Error("spooky!")
@@ -27,5 +27,5 @@ Error: spooky!
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:*)
     at Module.load (internal/modules/cjs/loader.js:*)
     at Function.Module._load (internal/modules/cjs/loader.js:*)
-    at Function.Module.runMain (internal/modules/cjs/loader.js:*)
+    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*)
     at internal/main/run_main_module.js:*:*
