+++ test/message/vm_dont_display_syntax_error.out
@@ -13,5 +13,5 @@ SyntaxError: Unexpected number
     at Object.Module._extensions..js (internal/modules/cjs/loader.js:*)
     at Module.load (internal/modules/cjs/loader.js:*)
     at Function.Module._load (internal/modules/cjs/loader.js:*)
-    at Function.Module.runMain (internal/modules/cjs/loader.js:*)
+    at Function.executeUserEntryPoint [as runMain] (internal/modules/run_main.js:*)
     at internal/main/run_main_module.js:*:*
